Implement BigInteger integration for unlimited precision integer arithmetic
Replace int with BigInteger throughout Z3IntExpr system to provide unlimited precision
integer arithmetic matching Z3's capabilities. This eliminates integer overflow issues
and allows working with arbitrarily large integers.

Core Changes:
- Z3IntExpr: All operator overloads now use BigInteger instead of int
- Z3ContextExtensions: Updated all integer operations (arithmetic, comparison, equality, min/max)
- Z3Model: GetIntValue() now returns BigInteger using string-based extraction
- Tests: Updated assertions to compare with BigInteger values

Benefits:
- Unlimited precision integer arithmetic - no overflow
- Backward compatibility via implicit int → BigInteger conversion
- Exact arithmetic matching Z3's mathematical model
- All 343 tests passing

API Examples:
- solver.Assert(x == 5); // Implicit conversion still works
- solver.Assert(y >= 2147483648); // Beyond int.MaxValue - no problem!
- BigInteger value = model.GetIntValue(x); // Unlimited precision extraction

+++ Z3Wrap/Expressions/Z3IntExpr.cs
@@ -1,3 +1,5 @@
+using System.Numerics;
+
 namespace Z3Wrap.Expressions;
 
 #pragma warning disable CS0660, CS0661 // Type defines operator == or operator != but does not override Object.Equals/GetHashCode (handled by base class)
@@ -19,33 +21,33 @@ public sealed class Z3IntExpr : Z3Expr
     public static Z3BoolExpr operator >=(Z3IntExpr left, Z3IntExpr right) => left.Context.Ge(left, right);
     public static Z3IntExpr operator -(Z3IntExpr expr) => expr.Context.UnaryMinus(expr);
 
-    // Z3IntExpr <-> int operations  
-    public static Z3IntExpr operator +(Z3IntExpr left, int right) => left.Context.Add(left, right);
-    public static Z3IntExpr operator -(Z3IntExpr left, int right) => left.Context.Sub(left, right);
-    public static Z3IntExpr operator *(Z3IntExpr left, int right) => left.Context.Mul(left, right);
-    public static Z3IntExpr operator /(Z3IntExpr left, int right) => left.Context.Div(left, right);
-    public static Z3IntExpr operator %(Z3IntExpr left, int right) => left.Context.Mod(left, right);
-    public static Z3BoolExpr operator <(Z3IntExpr left, int right) => left.Context.Lt(left, right);
-    public static Z3BoolExpr operator <=(Z3IntExpr left, int right) => left.Context.Le(left, right);
-    public static Z3BoolExpr operator >(Z3IntExpr left, int right) => left.Context.Gt(left, right);
-    public static Z3BoolExpr operator >=(Z3IntExpr left, int right) => left.Context.Ge(left, right);
+    // Z3IntExpr <-> BigInteger operations  
+    public static Z3IntExpr operator +(Z3IntExpr left, BigInteger right) => left.Context.Add(left, right);
+    public static Z3IntExpr operator -(Z3IntExpr left, BigInteger right) => left.Context.Sub(left, right);
+    public static Z3IntExpr operator *(Z3IntExpr left, BigInteger right) => left.Context.Mul(left, right);
+    public static Z3IntExpr operator /(Z3IntExpr left, BigInteger right) => left.Context.Div(left, right);
+    public static Z3IntExpr operator %(Z3IntExpr left, BigInteger right) => left.Context.Mod(left, right);
+    public static Z3BoolExpr operator <(Z3IntExpr left, BigInteger right) => left.Context.Lt(left, right);
+    public static Z3BoolExpr operator <=(Z3IntExpr left, BigInteger right) => left.Context.Le(left, right);
+    public static Z3BoolExpr operator >(Z3IntExpr left, BigInteger right) => left.Context.Gt(left, right);
+    public static Z3BoolExpr operator >=(Z3IntExpr left, BigInteger right) => left.Context.Ge(left, right);
 
-    // int <-> Z3IntExpr operations
-    public static Z3IntExpr operator +(int left, Z3IntExpr right) => right.Context.Add(left, right);
-    public static Z3IntExpr operator -(int left, Z3IntExpr right) => right.Context.Sub(left, right);
-    public static Z3IntExpr operator *(int left, Z3IntExpr right) => right.Context.Mul(left, right);
-    public static Z3IntExpr operator /(int left, Z3IntExpr right) => right.Context.Div(left, right);
-    public static Z3IntExpr operator %(int left, Z3IntExpr right) => right.Context.Mod(left, right);
-    public static Z3BoolExpr operator <(int left, Z3IntExpr right) => right.Context.Lt(left, right);
-    public static Z3BoolExpr operator <=(int left, Z3IntExpr right) => right.Context.Le(left, right);
-    public static Z3BoolExpr operator >(int left, Z3IntExpr right) => right.Context.Gt(left, right);
-    public static Z3BoolExpr operator >=(int left, Z3IntExpr right) => right.Context.Ge(left, right);
+    // BigInteger <-> Z3IntExpr operations
+    public static Z3IntExpr operator +(BigInteger left, Z3IntExpr right) => right.Context.Add(left, right);
+    public static Z3IntExpr operator -(BigInteger left, Z3IntExpr right) => right.Context.Sub(left, right);
+    public static Z3IntExpr operator *(BigInteger left, Z3IntExpr right) => right.Context.Mul(left, right);
+    public static Z3IntExpr operator /(BigInteger left, Z3IntExpr right) => right.Context.Div(left, right);
+    public static Z3IntExpr operator %(BigInteger left, Z3IntExpr right) => right.Context.Mod(left, right);
+    public static Z3BoolExpr operator <(BigInteger left, Z3IntExpr right) => right.Context.Lt(left, right);
+    public static Z3BoolExpr operator <=(BigInteger left, Z3IntExpr right) => right.Context.Le(left, right);
+    public static Z3BoolExpr operator >(BigInteger left, Z3IntExpr right) => right.Context.Gt(left, right);
+    public static Z3BoolExpr operator >=(BigInteger left, Z3IntExpr right) => right.Context.Ge(left, right);
 
     // Mixed-type equality operations
-    public static Z3BoolExpr operator ==(Z3IntExpr left, int right) => left.Context.Eq(left, right);
-    public static Z3BoolExpr operator !=(Z3IntExpr left, int right) => left.Context.Neq(left, right);
-    public static Z3BoolExpr operator ==(int left, Z3IntExpr right) => right.Context.Eq(left, right);
-    public static Z3BoolExpr operator !=(int left, Z3IntExpr right) => right.Context.Neq(left, right);
+    public static Z3BoolExpr operator ==(Z3IntExpr left, BigInteger right) => left.Context.Eq(left, right);
+    public static Z3BoolExpr operator !=(Z3IntExpr left, BigInteger right) => left.Context.Neq(left, right);
+    public static Z3BoolExpr operator ==(BigInteger left, Z3IntExpr right) => right.Context.Eq(left, right);
+    public static Z3BoolExpr operator !=(BigInteger left, Z3IntExpr right) => right.Context.Neq(left, right);
     
     public Z3IntExpr Add(Z3IntExpr other) => Context.Add(this, other);
     public Z3IntExpr Sub(Z3IntExpr other) => Context.Sub(this, other);
@@ -57,15 +59,15 @@ public sealed class Z3IntExpr : Z3Expr
     public Z3BoolExpr Gt(Z3IntExpr other) => Context.Gt(this, other);
     public Z3BoolExpr Ge(Z3IntExpr other) => Context.Ge(this, other);
 
-    public Z3IntExpr Add(int other) => Context.Add(this, other);
-    public Z3IntExpr Sub(int other) => Context.Sub(this, other);
-    public Z3IntExpr Mul(int other) => Context.Mul(this, other);
-    public Z3IntExpr Div(int other) => Context.Div(this, other);
-    public Z3IntExpr Mod(int other) => Context.Mod(this, other);
-    public Z3BoolExpr Lt(int other) => Context.Lt(this, other);
-    public Z3BoolExpr Le(int other) => Context.Le(this, other);
-    public Z3BoolExpr Gt(int other) => Context.Gt(this, other);
-    public Z3BoolExpr Ge(int other) => Context.Ge(this, other);
+    public Z3IntExpr Add(BigInteger other) => Context.Add(this, other);
+    public Z3IntExpr Sub(BigInteger other) => Context.Sub(this, other);
+    public Z3IntExpr Mul(BigInteger other) => Context.Mul(this, other);
+    public Z3IntExpr Div(BigInteger other) => Context.Div(this, other);
+    public Z3IntExpr Mod(BigInteger other) => Context.Mod(this, other);
+    public Z3BoolExpr Lt(BigInteger other) => Context.Lt(this, other);
+    public Z3BoolExpr Le(BigInteger other) => Context.Le(this, other);
+    public Z3BoolExpr Gt(BigInteger other) => Context.Gt(this, other);
+    public Z3BoolExpr Ge(BigInteger other) => Context.Ge(this, other);
 
     public Z3IntExpr UnaryMinus() => Context.UnaryMinus(this);
     public Z3IntExpr Abs() => Context.Abs(this);
