+++ Z3Wrap/Z3ContextExtensions.Comparison.cs
@@ -1,3 +1,4 @@
+using System.Numerics;
 using Z3Wrap.Expressions;
 using Z3Wrap.Interop;
 
@@ -30,15 +31,15 @@ public static partial class Z3ContextExtensions
         return context.WrapBoolExpr(resultHandle);
     }
 
-    // Z3IntExpr <-> int operations
-    public static Z3BoolExpr Lt(this Z3Context context, Z3IntExpr left, int right) => context.Lt(left, context.Int(right));
-    public static Z3BoolExpr Lt(this Z3Context context, int left, Z3IntExpr right) => context.Lt(context.Int(left), right);
-    public static Z3BoolExpr Le(this Z3Context context, Z3IntExpr left, int right) => context.Le(left, context.Int(right));
-    public static Z3BoolExpr Le(this Z3Context context, int left, Z3IntExpr right) => context.Le(context.Int(left), right);
-    public static Z3BoolExpr Gt(this Z3Context context, Z3IntExpr left, int right) => context.Gt(left, context.Int(right));
-    public static Z3BoolExpr Gt(this Z3Context context, int left, Z3IntExpr right) => context.Gt(context.Int(left), right);
-    public static Z3BoolExpr Ge(this Z3Context context, Z3IntExpr left, int right) => context.Ge(left, context.Int(right));
-    public static Z3BoolExpr Ge(this Z3Context context, int left, Z3IntExpr right) => context.Ge(context.Int(left), right);
+    // Z3IntExpr <-> BigInteger operations
+    public static Z3BoolExpr Lt(this Z3Context context, Z3IntExpr left, BigInteger right) => context.Lt(left, context.Int(right));
+    public static Z3BoolExpr Lt(this Z3Context context, BigInteger left, Z3IntExpr right) => context.Lt(context.Int(left), right);
+    public static Z3BoolExpr Le(this Z3Context context, Z3IntExpr left, BigInteger right) => context.Le(left, context.Int(right));
+    public static Z3BoolExpr Le(this Z3Context context, BigInteger left, Z3IntExpr right) => context.Le(context.Int(left), right);
+    public static Z3BoolExpr Gt(this Z3Context context, Z3IntExpr left, BigInteger right) => context.Gt(left, context.Int(right));
+    public static Z3BoolExpr Gt(this Z3Context context, BigInteger left, Z3IntExpr right) => context.Gt(context.Int(left), right);
+    public static Z3BoolExpr Ge(this Z3Context context, Z3IntExpr left, BigInteger right) => context.Ge(left, context.Int(right));
+    public static Z3BoolExpr Ge(this Z3Context context, BigInteger left, Z3IntExpr right) => context.Ge(context.Int(left), right);
 
     // Z3RealExpr <-> Z3RealExpr operations
     public static Z3BoolExpr Lt(this Z3Context context, Z3RealExpr left, Z3RealExpr right)
